rofl

+++ src/Darryldecode/Cart/Cart.php
@@ -705,7 +705,7 @@ class Cart {
     {
         if( preg_match('/\-/', $value) == 1 )
         {
-            $value = (int) str_replace('-','',$value);
+            $value = (float) str_replace('-','',$value);
 
             // we will not allowed to reduced quantity to 0, so if the given value
             // would result to item quantity of 0, we will not do it.
@@ -716,11 +716,11 @@ class Cart {
         }
         elseif( preg_match('/\+/', $value) == 1 )
         {
-            $item[$key] += (int) str_replace('+','',$value);
+            $item[$key] += (float) str_replace('+','',$value);
         }
         else
         {
-            $item[$key] += (int) $value;
+            $item[$key] += (float) $value;
         }
 
         return $item;
@@ -736,7 +736,7 @@ class Cart {
      */
     protected function updateQuantityNotRelative($item, $key, $value)
     {
-        $item[$key] = (int) $value;
+        $item[$key] = (float) $value;
 
         return $item;
     }
